Provide an API to defer causality checks (#63)
`CausalCell` provides a structure that verifies that the accessing
thread has synchronized appropriately to be able to to access the cell.
However, there are algorithms where a thread must read
**optimistically** from a causal cell before it knows that it can
perform the read. In this case, the thread reads the bytes from the cell
to a temporary location then performs the necessary atomic operation to
verify access.

To model this, a new API is added to CausalCell that allows deferring
the causality check to a later time. `with_deferred` returns a
`CausalCheck` that can be used at a later time to assert the previous
access was correct.