Check for leaks (#73)
A first pass at adding some leak checking to loom + other improvements.

`Arc` is promoted to a full "object". This allows optimizing the behavior of
the mocked arc with the loom scheduler. Now, using a mocked `Arc` requires
less permutations as loom is able to "see" through it. Also, by promoting it to
an object, at the end of each execution, all arcs are checked to see if they were
correctly freed. In the event that they were not freed, the execution fails.

The waker behind `block_on` is updated to use a mocked `Arc`. This allows
catching leaking wakers.

Mocked versions of `std::alloc::{alloc, dealloc}` are added. These do not impact
the scheduler or permutations, but it allows tracking correct usage of
allocation & deallocation within the context of an execution. If an allocation leaks,
loom fail the execution.

A `loom::alloc::Track<T>` type is added to allow tracking arbitrary values an
 ensure that the drop handlers are called. This functions in a similar way as
allocation tracking.

Finally, `Notify` is improved to permute on spurious wake ups. When waiting on
a notify, loom will first try the execution without any spurious wake ups, then it
will try the execution with a spurious wake up.

`futures::block_on` is updated to remove the double poll simulating spurious
wake ups as doing so could hide bugs.