fixes Loom when using async/await (#102)
Loom uses its own thread implementation using generators. Unfortunately,
this means that thread-locals do not work when using a loom model as all
code runs on a single physical thread. Loom provides a mocked
thread_local API to work around this.

However, when using `async / await`, `std` uses an internal thread-local
to track the waker as it passes through the async block. If loom
preempts at this point, the context gets shared with the wrong thread,
resulting in SEGV.

This patch fixes this by wrapping the loom context switch with an async
block and polling the async block w/ a noop waker. Doing this forces the
real waker to be moved out of thread-local storage and to the stack,
which is persisted across the loom context switch.