add additional location tracking to `Arc`, `alloc`, and `mpsc` (#265)
## Motivation

Currently, loom's mock `Arc` will track the source code location where
the `Arc` was created. This is useful for diagnosing problems, but it
isn't always *enough* to make the output easy to understand. 

In particular, `LOOM_LOG=trace` will indicate where the `Arc` was
created, and will log ref count operations on that `Arc`. The creation
location is useful , but ref count operations aren't associated with a
location --- so a user reading the log cannot determine _where_ the
`Arc`'s reference count was incremented or decremented from. This makes
tracking leaks difficult.

Additionally, in the `LOOM_LOG` output, each `Arc` instance is
identified by its `Ref` (an index in the store of loom object states).
This permits tracking an individual `Arc`'s refcount lifecycle in the
logs. However, when a leak is detected, the panic message does not
include this index, so it's not possible to determine which particular
`Arc` was leaked. The location where the `Arc` was created is included,
but multiple `Arc`s may have originated from the same location.

## Solution

This branch improves `Arc` debuggability by adding location tracking to
operations that increment and decrement the `Arc`'s reference count. It
also adds the index in the object store to the panic message when a leak
is detected, so the user can now go back and look for all the events for
the `Arc` instance with that ID. This should make debugging `Arc` leaks
much easier.

Similarly, I've also added location tracking to Loom's `alloc` and
`mpsc` mocks, and modified their panic messages to print the store index
as well. 

Signed-off-by: Eliza Weisman <eliza@buoyant.io>