add a span for the current model's iteration (#267)
This commit adds a `tracing` span with the current model's iteration
that's entered every time an iteration begins. This should make it much
easier to distinguish individual iterations in the `LOOM_LOG` output.

This was supposed to be part of #266 (and I discussed it in the PR
description for that branch) but I forgot to push this commit. Whoops.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>