docs: fix broken rustdoc links (#273)
PR #151 added a CI job to try and build the documentation, failing on
RustDoc errors. It turned out there was a handful of broken links, so
that job is now failing. 

I meant to fix this as part of #151 but accidentally automerged it
before fixing the docs, since automerge wasn't configured to require the
docs build job to pass. Whoops.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>